+++ R/retry_download.R
@@ -1,13 +1,18 @@
-#' Retry an expression `ntimes` times waiting an increasing amount of time
-#' between tries, i.e. waiting for `Sys.sleep(i * sleep_mult)` seconds between
-#' each try. If `expr` fails for `ntimes` times the error will be thrown.
+#' @title Retry expression on failure
+#'
+#' @description
+#'
+#' `retry()` retries, upon failure, the evaluation of an expression `exp` for
+#' `ntimes` times waiting an increasing amount of time
+#' between tries, i.e., waiting for `Sys.sleep(i * sleep_mult)` seconds between
+#' each try `i`. If `expr` fails for `ntimes` times an error will be thrown.
 #'
 #' @param expr Expression to be evaluated.
 #'
 #' @param ntimes `integer(1)` with the number of times to try.
 #'
 #' @param sleep_mult `numeric(1)` multiplier to define the increasing waiting
-#'     time.
+#'     time (in seconds).
 #'
 #' @note
 #'
@@ -17,8 +22,8 @@
 #'
 #' @importFrom methods is
 #'
-#' @noRd
-.retry <- function(expr, ntimes = 3, sleep_mult = 0) {
+#' @export
+retry <- function(expr, ntimes = 5L, sleep_mult = 0L) {
     res <- NULL
     for (i in seq_len(ntimes)) {
         res <- suppressWarnings(tryCatch(expr, error = function(e) e))
@@ -32,5 +37,5 @@
 }
 
 .sleep_mult <- function() {
-    as.integer(getOption("metabolights.sleep_mult", default = 5))
+    as.integer(getOption("metabolights.sleep_mult", default = 7L))
 }