+++ tests/testthat/test_retry_download.R
@@ -1,16 +1,17 @@
-test_that(".retry works", {
+test_that("retry works", {
     a <- function() {
         if (sample(0:1, 1) == 0)
             stop("A, got a 0")
         1
     }
-
     set.seed(123)
-    expect_error(.retry(a()), "A, got a 0")
-    res <- .retry(a())
+    res <- retry(a(), ntimes = 5L)
     expect_equal(res, 1)
+    ## Failure
+    set.seed(123)
+    expect_error(retry(a(), ntimes = 3L), "A, got a 0")
 })
 
 test_that(".sleep_mult works", {
-    expect_equal(.sleep_mult(), 5L)
+    expect_equal(.sleep_mult(), 7L)
 })