+++ NEWS.md
@@ -1,3 +1,10 @@
+# MsBackendMetaboLights 1.5
+
+# MsBackendMetaboLights 1.5.1
+
+- Increase number of download retries and increase waiting time in between.
+- Export the `retry()` function.
+
 # MsBackendMetaboLights 1.3
 
 ## Changes in 1.3.2